diag_network: Adding tool to benchmark network
This tool is made for testing two kind of network load on a cluster :
- 1 to 1 network link
- all to all network links (aka storm mode)

The "1to1" mode is made for insuring that all links between hosts are
working properly. It makes a simple netpipe and reports bandwidth and
link integrity.

In the gnuplot mode is used, the 'diag_network.gnuplot' script can be
called with gnuplot to trace the result.

If the storm mode '-s' is engaged, the test will use the kash (from
kanif) shell to start simultaneously the benchmarking from all the nodes
to all the other nodes. On a 6 nodes cluster, each node will generate 5
streams to the other nodes leading to a total of 6*5 (30) streams.
This test is loading both nodes & network backend (switch). It will show
how the setup behave under a complete load.